listen: accept concurrency
Though some amount of concurrency (namely the ipfs handshake) was
nominally moved out of the hot-path of the `Accept` call-- the
setupConn function call was still causing it to happen because
the transport was setup-- which wrote to the connection before it
was fully setup.

Accept concurrency is not an infinite resource though-- we must be
careful to rate limit and make sure that malicious connection
attempts do not get the process stuck, or OOM it. It should only
slow it down.